Add handler functions per message type:
This allows us to participate in the full
DORA(discover,offer,request,ack) cycle.
The release handler is a noop but include for
clarity. Release is a noop because the design of
this repo is that all DHCP addresses are reservations
so a release doesn't put an IP back into a pool.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>